Added the ability to hijack connections (for Websockets)
During my web socket tests I realized that ProxyWriter
and BufferWriter didn't implement the http.Hijacker
interface.

Moreover, there wasn't sufficient debugging info in the logs
that told us which implementation of ResponseWriter was being used.

This information is necessary in the future when proxies
are injected and we need to quickly figure out where the gaps are.

I've added logs which emit the type of the implementer of an interface,
whenever it cannot be cast into a target interface.

I also implemented http.Hijacker on ProxyWriter and BufferWriter.