Change scheme for WS requests when backends might be registered as HTTP.
Because of the way backends might be registered to Vulcan,
and if a backend supports both http and web socket protocols,
it makes sense for the web socket forwarder to rewrite the
scheme correctly.

It IS possible to do this with multiple routes, but that's fickle
and difficult to maintain when not necessary.