Use the ProxyWriter in ServeHTTP to get code and length
I noticed that the pw isn't actually used in ServeHTTP, and we're getting "code: 0, length: 0" in the log output. 
This change passes pw to ServeHTTP, and it also instantiates it as a pointer to satisfy the http.ResponseWriter interface.

There were no unit tests to update for this change. I'll try to write a test for the streaming forwarder case sometime soon.