Added Hijack and CloseNotifier interfaces
When testing an upgrade to WebSocket connections,
I noticed that the buffer.bufferWriter ResponseWriter proxy
does not implement the Hijack or CloseNotify methods.

The Hijack method is of prominence here - it allows the websockets code
to hijack the connection and run the WS proxy inline. Without that
any Frontends would have trouble upgrading the connection in the forwarder.

The CloseNotifier is more for house-keeping. It allows http.Serve to terminate
in-flight connections to backends, when the client-side connection breaks.

That means if a browser closes a tab, or you "CTRL+C" out of
a wget/curl, it'll rapidly clean up the dependent connections to vulcan's
backend.

This is useful for resource saving under either long-lived or high-latency
 connections.