Re-introduce Header Rewriter in Websockets
Figured out why web sockets weren't working with the
rewriter.

Turns out, removing Hop-By-Hop headers was also removing
the Connection and Upgrade headers which are crucial
for the dial.

This change modifies the HeaderRewriter
to not remove Hop-By-Hop headers when it's a
WebSocket request.

This ensures all Websocket receipients still get
all the X-Forward* headers for better logging,
monitoring and decision-making.