Added a RequestRewriteListener
In many situations where we consume Oxy,
it is useful to have informaiton on the routing
decisions made by RR and Rebalancer, on a per-request
level.

I made a similar change to the Forwarder where there's
a UrlForwardingStateListener to be able to figure
out when a connection is attempted and when it ends.

Logs are useful, but they're not easily parsable
in-proc. And sometimes it can be useful to know
which backends were selected to use, and which ones
are unused (for whatever reason.)

Initially I had an info-level log (which I still think
is beneficial), but now at a DebugLevel, and instead
delegate the introspection/logging through
this listener instead.

Also added Unit tests to ensure it works and is
wired in correctly. All tests are passing.