Simplify fwd.go, and enable WSS (Websockets over TLS) (#73)
* Added Hijack and CloseNotifier interfaces

When testing an upgrade to WebSocket connections,
I noticed that the buffer.bufferWriter ResponseWriter proxy
does not implement the Hijack or CloseNotify methods.

The Hijack method is of prominence here - it allows the websockets code
to hijack the connection and run the WS proxy inline. Without that
any Frontends would have trouble upgrading the connection in the forwarder.

The CloseNotifier is more for house-keeping. It allows http.Serve to terminate
in-flight connections to backends, when the client-side connection breaks.

That means if a browser closes a tab, or you "CTRL+C" out of
a wget/curl, it'll rapidly clean up the dependent connections to vulcan's
backend.

This is useful for resource saving under either long-lived or high-latency
 connections.

* Don't write anything once connection has been  hijacked

* Added logs to Websocket proxying

* Only error when there is an error

* Web sockets use TLS

* Outgoing scheme must be wss to use tls.Dial

* Better WSS/WS dial

* Gofmt

* Consolidated all code against one forwarder with different Service methods

* Accidentally deleted DefaultTransport

* Re-introduce Header Rewriter in Websockets

Figured out why web sockets weren't working with the
rewriter.

Turns out, removing Hop-By-Hop headers was also removing
the Connection and Upgrade headers which are crucial
for the dial.

This change modifies the HeaderRewriter
to not remove Hop-By-Hop headers when it's a
WebSocket request.

This ensures all Websocket receipients still get
all the X-Forward* headers for better logging,
monitoring and decision-making.

* Missed a reference in the unit test to the renamed function.

* Added a RequestRewriteListener

In many situations where we consume Oxy,
it is useful to have informaiton on the routing
decisions made by RR and Rebalancer, on a per-request
level.

I made a similar change to the Forwarder where there's
a UrlForwardingStateListener to be able to figure
out when a connection is attempted and when it ends.

Logs are useful, but they're not easily parsable
in-proc. And sometimes it can be useful to know
which backends were selected to use, and which ones
are unused (for whatever reason.)

Initially I had an info-level log (which I still think
is beneficial), but now at a DebugLevel, and instead
delegate the introspection/logging through
this listener instead.

Also added Unit tests to ensure it works and is
wired in correctly. All tests are passing.