restore time.Duration and time.Time.
I replaced all clock.Durations with time.Durations and only exposed
clock.Time's with time.Times. This was done so as to not change the
external API.

Note: Because of how go type aliases work I'm not convinced this actually
changed the public interfaces. clock.Time and clock.Duration are type
aliases to the normal time package equivalents.